Here is a setup for TRPO
All I did was carry around the old policy and set of states so that instead of multiplying by logpi, I multiply by logpi-logpi_old. the set of old states are so that I can compute the KL divergence. It seemed to run better then policy iteration so hopefully its working.